+++ Module07/ex02/Array.ipp
@@ -6,7 +6,7 @@ Array<T>::Array() \
 {}
 
 template <typename T>
-Array<T>::Array(size_t n)
+Array<T>::Array(uint n)
 	: _size(n), _arr(new T[n]())
 {}
 
@@ -33,13 +33,13 @@ Array<T>	&Array<T>::operator=(const Array<T> &copy)
 }
 
 template<typename T>
-size_t Array<T>::size() const
+uint Array<T>::size() const
 {
 	return (this->_size);
 }
 
 template<typename T>
-T &Array<T>::operator[](size_t i) const
+T &Array<T>::operator[](uint i) const
 {
 	if (i >= this->_size)
 		throw OutOfBoundsException();
@@ -57,7 +57,7 @@ void	Array<T>::_deepCopy(const Array<T> &copy)
 {
 	this->_size = copy._size;
 	this->_arr = new T[this->_size];
-	for (size_t i = 0; i < this->_size; ++i)
+	for (uint i = 0; i < this->_size; ++i)
 		this->_arr[i] = copy._arr[i];
 }
